chore: Move job display logic to commands module
The Display trait implementations for Job and JobList were moved from
application.rs to a new display_jobs function in commands.rs, along with
their tests.
